requirements.txt
saxo-openapi>=0.8.1
requests>=2.32.0
urllib3>=2.2.2
pyyaml>=6.0
python-dotenv>=1.0.0
